refactor: modal을 domain 로직으로 지정
1. specific 디렉토리로 이동
2. shareModal -> spaceShareModal로 네이밍 변경
