+++ frontend/src/components/specific/modal/spaceShareModal/SpaceShareModal.styles.ts
